+++ package.js
@@ -34,6 +34,8 @@ function configure( api ){
     api.use( 'pwix:modal@1.10.0 || 2.0.0' );
     api.use( 'pwix:tabbed@1.0.0-rc' );
     api.use( 'pwix:ui-bootstrap5@2.0.0' );
+    api.use( 'pwix:ui-fontawesome6@1.0.0' );
+    api.use( 'pwix:ui-utils@1.0.0' );
     api.use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
     api.use( 'tracker' );
     api.addFiles( 'src/client/components/prEdit/prEdit.js', 'client' );