+++ src/common/js/startup.js
@@ -4,7 +4,5 @@
 
 // be verbose if asked for
 Meteor.startup( function(){
-    if( Roles.configure().verbosity & Roles.C.Verbose.CONFIGURE ){
-        console.debug( 'pwix:roles common.startup()', Roles );
-    }
+    _verbose( Roles.C.Verbose.STARTUP, 'pwix:roles common startup()' );
 });