+++ ChangeLog.md
@@ -9,8 +9,9 @@
     - Define Roles.usedScopes() new function, bumping minor candidate version number
     - Roles.configure() becomes a reactive data source
     - Introduce trace module
-    - Introduce 'pr-edit-state' new event
+    - Introduce 'pr-global-state' new event
     - Define Roles.scopedRoles() new function
+    - Add missing pwix:ui-utils and ui-fontawesome6 dependencies
 
 ### 1.4.0